+++ src/components/Footer/index.js
@@ -6,8 +6,7 @@ const Footer = () => {
       <div className="row">
         <div className="col-4">
           <p>
-            This demo store aims to showcase the functions of Voucherify. This
-            is not a real shop.
+            This demo store aims to showcase the functions of <a href='https://voucherify.io'>Voucherify</a>.
           </p>
           <small className="d-block mb-3 text-muted">
             © 2019-{new Date().getFullYear()}