+++ src/redux/reducers/index.js
@@ -1,10 +1,10 @@
 import {combineReducers} from 'redux';
 import {userReducer} from './userReducer'
 import {cartReducer} from './cartReducer'
-import {shopReducer} from './shopReducer'
+import {storeReducer} from './storeReducer'
 
 export default combineReducers({
   cartReducer,
   userReducer,
-  shopReducer,
+  storeReducer,
 });